add and remove item on wishlist,useCase classes for check for update wishlist

+++ app/src/main/java/com/rimapps/shoppingapp/features/productList/repository/api/ProductRepositoryApi.kt
@@ -15,10 +15,9 @@ class ProductRepositoryApi @Inject constructor(private val service: ProductServi
                 Product(
                     it.title,
                     it.description,
-                    it.price.toDouble(),
+                    it.price,
                     it.imageUrl,
                     it.id
-
                 )
             }
